MigrationFX004

+++ db/migrate/20250413173742_remove_photo_order_from_card_collections.rb
@@ -1,5 +1,7 @@
 class RemovePhotoOrderFromCardCollections < ActiveRecord::Migration[6.1]
   def change
-    remove_column :card_collections, :photo_order, :text
+    if column_exists?(:card_collections, :photo_order)
+      remove_column :card_collections, :photo_order, :text
+    end
   end
 end
